requirements.txt
python-telegram-bot==21.9
supabase==2.27.0
python-dotenv==1.0.0
